requirements.txt
fastapi==0.101.1     
fastapi_restful==0.5.0
pydantic==2.2.1
SQLAlchemy==2.0.20
starlette==0.27.0   
uvicorn==0.23.2     
mysqlclient==2.2.0
redis==5.0.0
celery==5.3.1        
selenium==4.11.2
kafka-python==2.0.2
python_dotenv==1.0.0
typing-inspect==0.9.0
pydantic-settings==2.0.3
urllib3==1.26.7
pytube==15.0.0